#0 Readme update

+++ src/core/GLObject.cpp
@@ -42,7 +42,7 @@ namespace core {
 
 GLObject::GLObject(GLsizei howMany) :
 	mHowMany(howMany) {
-	if (mHowMany > FILLWAVE_GLOBJECT_MAX_CAPACITY) {
+	if (mHowMany > static_cast<int>(FILLWAVE_GLOBJECT_MAX_CAPACITY)) {
 		FLOG_FATAL("Currect globject size exceeded");
 	}
